better output presentation of game states refactor static methods into Blackjack's methods

+++ src/com/ivanfoong/insightdataengineering/blackjack/user/Dealer.java
@@ -4,7 +4,7 @@ package com.ivanfoong.insightdataengineering.blackjack.user;
  * Created by ivanfoong on 4/4/14.
  */
 public class Dealer extends Person {
-    public Dealer(final Integer aTotalValue) {
+    public Dealer(final Double aTotalValue) {
         super(aTotalValue);
     }
 }
